[refactoring] Extract production code to web folder.
